requirements.txt
Flask==3.1.2
huggingface_hub==0.35.3
matplotlib==3.10.6
openai==2.1.0
pdfplumber==0.11.7
python-dotenv==1.1.1
